Moved mutable renderers to new file
